Feat: Navigation 바 추가 및 레이아웃 정리
